Now takes the optional numproc parameter for specifying the number of processors.

+++ optFDA.py
@@ -1,16 +1,17 @@
 import random, os, sys, time, shutil
 from tempfile import mkstemp
-from subprocess import call
+from subprocess import call, Popen, PIPE
 from random import Random
 import inspyred
+import ConfigParser
 
 """ 
     Copyright (c) 2013, Ergun Bicici <ergunbicici@yahoo.com>
     
     dmin, dmax, cmin, cmax, smin, smax, imin, imax, lmin, lmax values can be changed to select a different space for the optimization.
-
+    
     numberofprocessors determine the number of processors to use during optimization.
-
+    
 """
 
 dmin = 0.0001
@@ -25,21 +26,30 @@ imax = 5.0001
 lmin = -5.0001
 lmax = 5.0001
 
-numberofprocessors = 6
-tmpdir = 'tmp'
+numberofprocessors = 24
 
-def parseFDAScores(errfile):
-    f = open(errfile, 'r')
-    l = f.readline()
-    f.close()
-    os.remove(errfile)
+def parseFDAScores(errout):
+    if type(errout) == file:
+        f = open(errfile, 'r')
+        l = f.readline()
+        f.close()
+        os.remove(errfile)
+    else:
+        l = errout
     parts = l.split()
     tffound = float(parts[-1])
     sffound = float(parts[-4])
     alltf = float(parts[-2])
     allsf = float(parts[-5])
-    scov = sffound / allsf
-    tcov = tffound / alltf
+    try:
+        scov = sffound / allsf
+        tcov = tffound / alltf
+    except:
+        print errout
+        print parts
+        print sffound, allsf
+        print tffound, alltf
+        sys.exit()
     numsents = float(parts[-7])
     return (scov, tcov, allsf, alltf, sffound, tffound, numsents)
 
@@ -60,27 +70,18 @@ def bound_FDA(candidate, args):
     candidate = [d, c, s, i, l]
     return candidate
 
-def FDAEvalFunction(newguess): #, verbose=0):
+def FDAEvalFunction(newguess):
     global fda_path, train_source_path, train_target_path, test_source_path, test_target_path, fda_n, fda_numwords
     d = newguess[0]
     c = newguess[1]
     s = newguess[2]
     i = newguess[3]
     l = newguess[4]
-
+    
     traintest_str = train_source_path + ' ' + test_source_path + ' ' + train_target_path + ' ' + test_target_path
-    if not os.path.exists(tmpdir):
-        try:
-            os.mkdir(tmpdir)
-        except:
-            pass
-    (fd, errfname) = mkstemp(dir=tmpdir)
-    cmd = fda_path + ' -v1 -t' + str(fda_numwords) +' -d'+str(d) +' -c'+str(c) + ' -s'+str(s) + ' -i'+str(i)+ ' -l'+str(l)+ ' -n'+str(fda_n) + ' -o /dev/null ' + ' ' + traintest_str + ' 2> ' + errfname
-    call(cmd, shell=True)
-    (scov, tcov, allsf, alltf, sffound, tffound, numsents) = parseFDAScores(errfname)
-#    if verbose:
-#        print tcov, scov, '[', ' '.join(map(str, newguess)), ']', 'n:', fda_n
-#        sys.stdout.flush()
+    cmd = fda_path + ' -v1 -t' + str(fda_numwords) +' -d'+str(d) +' -c'+str(c) + ' -s'+str(s) + ' -i'+str(i)+ ' -l'+str(l)+ ' -n'+str(fda_n) + ' -o /dev/null ' + ' ' + traintest_str
+    errout = Popen(cmd, shell=True, stderr=PIPE).stderr.readlines()[0]
+    (scov, tcov, allsf, alltf, sffound, tffound, numsents) = parseFDAScores(errout)
     return tcov
 
 def evaluate_FDA5(candidates, args):
@@ -90,15 +91,20 @@ def evaluate_FDA5(candidates, args):
         fitness.append(fit)
     return fitness
 
-def main(configFile, prng=None, display=False):
+def main(numproc=20, prng=None, display=True):
+    global numberofprocessors
+    numberofprocessors = numproc
     if prng is None:
         prng = Random()
         prng.seed(time.time())
-
+    
     ea = inspyred.ec.ES(prng)
     ea.terminator = [inspyred.ec.terminators.evaluation_termination,
                      inspyred.ec.terminators.diversity_termination,
                      inspyred.ec.terminators.generation_termination]
+    ea.observer = [inspyred.ec.observers.stats_observer,
+                   inspyred.ec.observers.best_observer,
+                   inspyred.ec.observers.file_observer]
     final_pop = ea.evolve(generator = generate_randomFDA5Params, 
                           evaluator = inspyred.ec.evaluators.parallel_evaluation_mp,
                           mp_evaluator = evaluate_FDA5, 
@@ -107,11 +113,11 @@ def main(configFile, prng=None, display=False):
                           bounder = bound_FDA,
                           maximize = True,
                           max_evaluations = 10000,
-                          max_generations = 5,
+                          max_generations = 10,
                           num_inputs=5)
     
     if display:
-        best = max(final_pop) 
+        best = max(final_pop)
         print('Best Solution: \n{0}'.format(str(best)))
     """
     The first n items are the solution, the second n items are the strategy parameters:
@@ -124,8 +130,22 @@ def main(configFile, prng=None, display=False):
     also the strategy parameters), so normal evaluator functions may be used
     seamlessly.
     """
-    if os.path.exists(tmpdir):
-        shutil.rmtree(tmpdir)
+
+def runConfig(configfile, numproc):
+    global numberofprocessors
+    global fda_path, train_source_path, train_target_path, test_source_path, test_target_path, fda_n, fda_numwords
+    numberofprocessors = numproc
+    config = ConfigParser.SafeConfigParser()
+    config.read(configfile)
+    # SMTPar: SMT Parameters
+    fda_path = config.get('FDAPar', 'fda_path')
+    train_source_path = config.get('FDAPar', 'train_source')
+    train_target_path = config.get('FDAPar', 'train_target')
+    test_source_path = config.get('FDAPar', 'test_source')
+    test_target_path = config.get('FDAPar', 'test_target')
+    fda_n = config.getint('FDAPar', 'n')
+    fda_numwords = config.getint('FDAPar', 'numwords')
+    main()
 
 def getCommands():
     import ConfigParser
@@ -135,6 +155,8 @@ def getCommands():
     usage = "usage: %prog [options] arg1 arg2 ..."
     parser = OptionParser(usage=usage, version="%prog 1.0")
     parser.add_option("-c", "--configfile", dest="configfile",
+                      help="supply a user defined configfile")
+    parser.add_option("-p", "--numproc", dest="numproc",
                       help="supply a user defined configfile")
     (options, args) = parser.parse_args()
     
@@ -155,6 +177,7 @@ def getCommands():
             config.read(options.configfile)
             
             # SMTPar: SMT Parameters
+            numproc = int(config.get('FDAPar', 'numproc'))
             fda_path = config.get('FDAPar', 'fda_path')
             train_source_path = config.get('FDAPar', 'train_source')
             train_target_path = config.get('FDAPar', 'train_target')
@@ -162,7 +185,7 @@ def getCommands():
             test_target_path = config.get('FDAPar', 'test_target')
             fda_n = config.getint('FDAPar', 'n')
             fda_numwords = config.getint('FDAPar', 'numwords')
-#            print fda_path, train_source_path, train_target_path, test_source_path, test_target_path, fda_n, fda_numwords
+            print fda_path, train_source_path, train_target_path, test_source_path, test_target_path, fda_n, fda_numwords
             if not os.path.exists(fda_path):
                 print fda_path, 'does not exist.'
                 sys.exit()
@@ -178,7 +201,11 @@ def getCommands():
             if not os.path.exists(test_target_path):
                 print test_target_path, 'does not exist.'
                 sys.exit()
-            main(options.configfile, display=True)
+            if options.numproc:
+                if int(options.numproc) != numproc:
+                    numproc = int(options.numproc)
+            main(numproc=numproc, display=True)
 
 if __name__ == '__main__':
     getCommands()
+
